django projects/blog django/requirements.txt
asgiref==3.5.2
attrs==21.4.0
backports.zoneinfo==0.2.1
boto3==1.24.16
botocore==1.27.16
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==3.2
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-filter==22.1
django-rest-framework==0.1.0
django-storages==1.12.3
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
idna==3.3
importlib-resources==5.9.0
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.7.2
oauthlib==3.2.0
Pillow==9.1.1
pycparser==2.21
PyJWT==2.4.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
zipp==3.8.1
gunicorn==20.1.0
psycopg2==2.9.3